+++ hypervisor/Makefile
@@ -95,14 +95,13 @@ D_SRCS += $(wildcard debug/*.c)
 C_SRCS += boot/acpi.c
 C_SRCS += boot/dmar_parse.c
 C_SRCS += arch/x86/ioapic.c
-C_SRCS += arch/x86/intr_lapic.c
+C_SRCS += arch/x86/lapic.c
 S_SRCS += arch/x86/trampoline.S
 C_SRCS += arch/x86/cpu.c
 C_SRCS += arch/x86/softirq.c
 C_SRCS += arch/x86/cpuid.c
 C_SRCS += arch/x86/mmu.c
 C_SRCS += arch/x86/notify.c
-C_SRCS += arch/x86/intr_main.c
 C_SRCS += arch/x86/vtd.c
 C_SRCS += arch/x86/gdt.c
 S_SRCS += arch/x86/cpu_primary.S
@@ -112,7 +111,7 @@ C_SRCS += arch/x86/timer.c
 C_SRCS += arch/x86/ept.c
 S_SRCS += arch/x86/vmx_asm.S
 C_SRCS += arch/x86/io.c
-C_SRCS += arch/x86/interrupt.c
+C_SRCS += arch/x86/virq.c
 C_SRCS += arch/x86/vmexit.c
 C_SRCS += arch/x86/vmx.c
 C_SRCS += arch/x86/assign.c
