+++ hypervisor/arch/x86/irq.c
@@ -732,3 +732,27 @@ void get_cpu_interrupt_info(char *str, int str_max)
 	}
 	snprintf(str, size, "\r\n");
 }
+
+int interrupt_init(uint32_t cpu_id)
+{
+	struct host_idt_descriptor *idtd = &HOST_IDTR;
+	int status;
+
+	set_idt(idtd);
+
+	status = init_lapic(cpu_id);
+	ASSERT(status == 0, "lapic init failed");
+	if (status != 0)
+		return -ENODEV;
+
+	status = init_default_irqs(cpu_id);
+	ASSERT(status == 0, "irqs init failed");
+	if (status != 0)
+		return -ENODEV;
+
+#ifndef CONFIG_EFI_STUB
+	CPU_IRQ_ENABLE();
+#endif
+
+	return status;
+}
