+++ hypervisor/arch/x86/lapic.c
